requirements.txt
requests>=2.28.0
colorama>=0.4.4
numpy==1.24.3
torch==2.1.1
transformers==4.35.2
python-dotenv==1.0.0
PyGithub==2.1.1
tqdm==4.66.1